Move TestParseSupport.wrap() to ParseSupport
