+++ presto-system-access-control/src/main/java/com/facebook/presto/security/ParseSupport.java
@@ -93,4 +93,9 @@ class ParseSupport
         }
         return -1;
     }
+
+    public static CharBuffer wrap(String s)
+    {
+        return CharBuffer.wrap(s).asReadOnlyBuffer();
+    }
 }
